+++ hazelcast-sql/src/test/java/com/hazelcast/sql/impl/calcite/validate/HazelcastSqlOperatorTableTest.java
@@ -18,6 +18,8 @@ package com.hazelcast.sql.impl.calcite.validate;
 
 import com.hazelcast.internal.util.BiTuple;
 import com.hazelcast.sql.impl.calcite.validate.operators.HazelcastReturnTypeInference;
+import com.hazelcast.sql.impl.calcite.validate.operators.common.AbstractCaseOperator;
+import com.hazelcast.sql.impl.calcite.validate.operators.common.HazelcastBinaryOperator;
 import com.hazelcast.sql.impl.calcite.validate.operators.common.HazelcastFunction;
 import com.hazelcast.sql.impl.calcite.validate.operators.common.HazelcastOperandTypeCheckerAware;
 import com.hazelcast.test.HazelcastParallelClassRunner;
