+++ hazelcast-sql/src/test/java/com/hazelcast/sql/impl/expression/NestingAndCasingExpressionTest.java
@@ -443,6 +443,11 @@ public class NestingAndCasingExpressionTest extends ExpressionTestSupport {
         check(sqlWithWhere("NOT BETWEEN SYMMETRIC ? AND ? "), SqlColumnType.INTEGER, 2, 2);
     }
 
+    @Test
+    public void test_CASE() {
+        check(sql("CASE ? WHEN this THEN '100' END"), 1);
+    }
+
     private void check(String sql, Object... params) {
         checkValue0(sql, SqlColumnType.VARCHAR, SKIP_VALUE_CHECK, params);
         checkValue0(sql.toLowerCase(), SqlColumnType.VARCHAR, SKIP_VALUE_CHECK, params);
